Putting all IBM Model M projects under one folder
As discussed in #12370:
The maintainers of several Model M replacement
controller projects agreed to cosolidate them
under one common folder.
