--- keyboards/handwired/ibm122m/readme.md
@@ -1,16 +1,0 @@
-# ibm122m
-
-![IBM Model M 122 key](https://i.imgur.com/Oo3Ozqz.jpg)
-
-This is a keymap for the IBM Model M 122 key terminal keyboard running on a Teensy 2.0++
-I wired it to weird pins on mine (mainly to accomodate speakers), so make sure to update the pin arrays.
-
-Keyboard Maintainer: [Luke Stanley](https://github.com/lukexorz)  
-Hardware Supported: Teensy 2.0++  
-Hardware Availability: https://www.pjrc.com/store/teensypp.html 
-
-Make example for this keyboard (after setting up your build environment):
-
-    make handwired/ibm122m:default
-
-See [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) then the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information.
