+++ keyboards/modelm/ashpil_usbc/ashpil_usbc.c
@@ -1,4 +1,4 @@
-/* Copyright 2019 iw0rm3r
+/* Copyright 2019 ashpil
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -11,36 +11,25 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <http://www./gnu.org/licenses/>.
  */
-#include "modelm101.h"
+#include "ashpil_usbc.h"
 
 void keyboard_pre_init_kb(void) {
   /* Setting status LEDs pins to output and +5V (off) */
-  setPinOutput(B4);
-  setPinOutput(B5);
-  setPinOutput(B6);
-  writePinHigh(B4);
-  writePinHigh(B5);
-  writePinHigh(B6);
+  setPinOutput(D5);
+  setPinOutput(D6);
+  setPinOutput(D7);
+  writePinHigh(D5);
+  writePinHigh(D6);
+  writePinHigh(D7);
 }
 
-void led_set_kb(uint8_t usb_led) {
-  if (usb_led & (1<<USB_LED_NUM_LOCK)) {
-    writePinLow(B4);
-  } else {
-    writePinHigh(B4);
-  }
-  if (usb_led & (1<<USB_LED_CAPS_LOCK)) {
-    writePinLow(B6);
-  } else {
-    writePinHigh(B6);
-  }
-  if (usb_led & (1<<USB_LED_SCROLL_LOCK)) {
-    writePinLow(B5);
-  } else {
-    writePinHigh(B5);
-  }
-
-  led_set_user(usb_led);
+bool led_update_kb(led_t led_state) {
+    if(led_update_user(led_state)) {
+        writePin(D5, !led_state.num_lock);
+        writePin(D6, !led_state.caps_lock);
+        writePin(D7, !led_state.scroll_lock);
+    }
+    return true;
 }
