+++ keyboards/modelm/ashpil_usbc/readme.md
@@ -9,6 +9,6 @@ This is a configuration of QMK intended to be used with the [USB C Model M alt c
 
 Make example for this keyboard (after setting up your build environment):
 
-    make ashpil/modelm_usbc:default
+    make modelm/ashpil_usbc:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
