+++ keyboards/modelm/ibm122m/info.json
@@ -1,5 +1,6 @@
 {
   "keyboard_name": "IBM Model M 122-key",
+  "url": "https://github.com/qmk/qmk_firmware/tree/master/keyboards/modelm/ibm122m",
   "maintainer": "qmk",
   "layouts": {
     "LAYOUT": {
