+++ keyboards/modelm/ibm122m/keymaps/default/config.h
