+++ keyboards/modelm/readme.md
@@ -1,0 +1,23 @@
+# IBM Model M
+
+This is the parent directory for all projects that replace the controller
+of IBM's Model M Keyboard and then run QMK on it.
+
+These controllers differ in their use of microcontrollers, connection (mainly
+USB) optional speakers etc.
+
+Some of them can be bought fully assembled others are do-it-yourself.
+
+## Warning - Model M Variants
+
+**Not all of these projects will fit into your keyboard!**
+
+As the Model M was produced over several years there are many variants of the
+controller PCB. Check what fits your keyboard before you build/order one.
+
+The variants built for PCs differ in:
+* PCB size
+* Connectors (8/12 pin flex, location)
+* Ground wire connector location
+
+Additionally IBM built Model M keyboards for terminals with 122 keys etc.
