+++ keyboards/modelm/yugo_m/readme.md
@@ -12,6 +12,6 @@ A drop-in replacement controller PCB for IBM Model M keyboards, based on the STM
 
 Make example for this keyboard (after setting up your build environment):
 
-    make yugo_m/model_m_101:default
+    make modelm/yugo_m:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
