+++ keyboards/modelm/yugo_m/yugo_m.c
@@ -1,4 +1,4 @@
-/* Copyright 2019 ashpil
+/* Copyright 2020 Nidzo Tomic <tomicn8@hotmail.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -11,25 +11,25 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www./gnu.org/licenses/>.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "modelm_usbc.h"
+#include "yugo_m.h"
 
 void keyboard_pre_init_kb(void) {
-  /* Setting status LEDs pins to output and +5V (off) */
-  setPinOutput(D5);
-  setPinOutput(D6);
-  setPinOutput(D7);
-  writePinHigh(D5);
-  writePinHigh(D6);
-  writePinHigh(D7);
+  // Set our LED pins as output
+  setPinOutput(A2);
+  setPinOutput(A1);
+  setPinOutput(A0);
+
+  keyboard_pre_init_user();
 }
 
 bool led_update_kb(led_t led_state) {
-    if(led_update_user(led_state)) {
-        writePin(D5, !led_state.num_lock);
-        writePin(D6, !led_state.caps_lock);
-        writePin(D7, !led_state.scroll_lock);
+    bool res = led_update_user(led_state);
+    if(res) {
+        writePin(A2, !led_state.num_lock);
+        writePin(A1, !led_state.caps_lock);
+        writePin(A0, !led_state.scroll_lock);
     }
-    return true;
+    return res;
 }
