--- keyboards/yugo_m/readme.md
@@ -1,11 +1,0 @@
-# Yugo-M - An STM32-based IBM Model M controller
-
-A drop-in replacement controller PCB for IBM Model M keyboards. 
-Currently available only for 101/102 key keyboards. 
-
-* Keyboard Maintainer: [Nidzo Tomic](https://github.com/tomic1785)
-* Hardware Supported: Yugo-M PCB
-  * Rev 0.9b
-  * Rev 1.1b
-
-For more info visit the project repository: [Yugo-M-controller-project](https://github.com/tomic1785/Yugo-M-controller-project)
